+++ drive/applications/application.hpp
@@ -9,7 +9,7 @@
 #include <libhal/steady_clock.hpp>
 #include <libhal/timeout.hpp>
 
-#include "../platform-implementations/ackermann_steering.hpp"
+#include "../include/ackermann_steering.hpp"
 
 namespace sjsu::drive {
